add: specific js in each pages
