--- dashboard/templates/dashboard/index.html
@@ -1,16 +1,0 @@
-{% extends "partials/base.html" %}
-{% block title %}
-AssetTrack Overview
-{% endblock %}
-{% block main %}
-<!-- Content Sections -->
-    <main id="mainSection" class="p-6 flex-1">
-        <div id="pageSection" class="section-content fade-in">
-            {% block page %} {% endblock %}
-        </div>
-    </main>   
-{% endblock %}
-{% block modals %}
-    <div id="formModal" class="fixed inset-0 bg-black/50 z-50 hidden flex items-center justify-center p-4">
-    </div>
-{% endblock %}
