+++ dashboard/templates/dashboard/sections/assets.html
@@ -1,6 +1,7 @@
-{% extends 'dashboard/index.html' %}
+{% extends 'partials/base.html' %}
+{% load static %}
+{% block title %} Assets {% endblock %}
 {% block page %}
-
     <div class="mb-6">
         <h3 class="text-xl font-semibold text-gray-800">Asset Catalog</h3>
         <p class="text-gray-600">View and manage all assets in your inventory</p>
@@ -65,4 +66,7 @@
             </table>
         </div>
     </div>
+{% endblock %}
+{% block scripts %} 
+    <script type="module" src="{% static 'js/asset.js' %}"></script>
 {% endblock %}
