+++ dashboard/templates/dashboard/sections/orders.html
@@ -1,4 +1,6 @@
-{% extends 'dashboard/index.html' %}
+{% extends 'partials/base.html' %}
+{% load static %}
+{% block title %} Orders {% endblock %}
 {% block page %}
     <div class="mb-6">
         <h3 class="text-xl font-semibold text-gray-800">Order Management</h3>
@@ -136,4 +138,7 @@
             </table>
         </div>
     </div>
+{% endblock %}
+{% block scripts %} 
+    <script type="module" src="{% static 'js/order.js' %}"></script>
 {% endblock %}
