+++ dashboard/templates/dashboard/sections/staff.html
@@ -1,4 +1,6 @@
-{% extends 'dashboard/index.html' %}
+{% extends 'partials/base.html' %}
+{% load static %}
+{% block title %} Staff {% endblock %}
 {% block page %}
     <div class="mb-6">
         <h3 class="text-xl font-semibold text-gray-800">Staff Management</h3>
@@ -152,4 +154,7 @@
             </table>
         </div>
     </div>
+{% endblock %}
+{% block scripts %} 
+    <script type="module" src="{% static 'js/staff.js' %}"></script>
 {% endblock %}
