+++ static/js/asset.js
@@ -1,0 +1,5 @@
+import { showSection } from './global.js';
+
+document.addEventListener('DOMContentLoaded', () => {
+    showSection('assets')
+})
